+++ utils_test.go
@@ -6,7 +6,7 @@ import (
 )
 
 func BenchmarkHeight(b *testing.B) {
-	h := int16(0)
+	h := uint32(0)
 	x := uint64(rand.Int63())
 	z, l := uint32(rand.Int31()), uint32(rand.Int31())
